Remove Stripe/Twilio basic auth criticism
The author is confusing two unrelated topics (unencrypted credentials
and header vs query param).

It's true that it's possible to accidentally expose your credentials if
they are sent unencrypted as an HTTP Authorization header. It is just as
easy to send them unencrypted as query params.

It's true that you can solve this by signing your requests, preferably
with an HMAC. You can put this signature in the HTTP Authorization
header to avoid accidental exposure too.

Basic auth is completely unrelated to the problem of combining plaintext
and using the secrets directly.