Remove XSS/CSRF paragraph
1. This is probably not the place to bring it up. Specific
vulnerabilities should probably be discussed after the actual
implementations.
2. In a XSS or CSRF attack, it's not the cookies that are nefarious or
insecure. The browser's policy to expose cookies to JS and send cookies
on all requests is the issue.
3. The `secure` flag is not a mitigation for CSRF. It's rarely used as a
mitigation for XSS since this means you can't access the cookies in JS
and there are many other things you can do with arbitrary JS execution
besides stealing the auth cookies. Even if it was, this is probably not
the place to present mitigations of vulnerabilities of implementations
we haven't discussed yet.